% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{nougad.benchmark}
\alias{nougad.benchmark}
\title{Benchmark the unmixing, print some interesting information, plot out some
plots}
\usage{
nougad.benchmark(
  n = 10240,
  k = 30,
  d = 40,
  rnw = 3,
  rpw = 1,
  nw = 5,
  rnwf = function(sp) rnw,
  rpwf = function(sp) rpw,
  nwf = function(sp) nw,
  iters = 250L,
  alpha = 0.02,
  accel = 1,
  threads = 0L,
  n_pheno = 20,
  pheno_positive_probability = 0.2,
  spectra_sdev = 0.2,
  spectra_intensity = 4,
  positive_population_intensity = 4,
  population_intensity_sdev = 0.1,
  emission_rate = 0.999,
  emission_quantum = 10,
  crosstalk_sdev = 5e-04,
  asinh_cofactor = 100,
  plotf = plot
)
}
\description{
Benchmark the unmixing, print some interesting information, plot out some
plots
}
