% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{nougad.benchmark}
\alias{nougad.benchmark}
\title{Benchmark the unmixing, print some interesting information, plot out some
plots}
\usage{
nougad.benchmark(
  n = 10240,
  k = 30,
  d = 40,
  rnw = 3,
  rpw = 1,
  nw = 1,
  rnwf = function(sp) rnw,
  rpwf = function(sp) rpw,
  nwf = function(sp) nw,
  alpha = 0.05,
  accel = 1,
  iters = 250L,
  n_pheno = 20,
  pheno_positive_probability = 0.2,
  plot.k = min(c(k, 5)),
  plotf = plot
)
}
\description{
Benchmark the unmixing, print some interesting information, plot out some
plots
}
